requirements.txt
tensorflow-gpu
tflite_support
pillow
numpy
tensorflow_datasets